refactor(memory): extract SpectorMemoryFactory for subsystem construction
- Extracted constructor logic from DefaultSpectorMemory to SpectorMemoryFactory.assemble()
- SubsystemBundle holds and passes the configured fields to DefaultSpectorMemory
- Cleaned up rebuildHnswIfNeeded by moving it to the factory as a static method
- Kept JVM Shutdown hook registration inside DefaultSpectorMemory to avoid dependency cycle
