+++ spector-memory/src/main/java/com/spectrayan/spector/memory/DefaultSpectorMemory.java
@@ -220,437 +220,46 @@ public final class DefaultSpectorMemory implements SpectorMemory, SpectorMemoryA
     private final com.spectrayan.spector.memory.pipeline.AttachmentProcessor attachmentProcessor;
 
     DefaultSpectorMemory(SpectorMemoryBuilder builder) {
+        var bundle = SpectorMemoryFactory.assemble(builder);
+        this.cognitiveTarget = bundle.cognitiveTarget();
+        this.embeddingProvider = bundle.embeddingProvider();
+        this.recallPipeline = bundle.recallPipeline();
+        this.index = bundle.index();
+        this.quantizer = bundle.quantizer();
+        this.partitionManager = bundle.partitionManager();
+        this.importanceEstimator = bundle.importanceEstimator();
+        this.reflectionOrchestrator = bundle.reflectionOrchestrator();
+        this.reinforcementHandler = bundle.reinforcementHandler();
+        this.valenceTracker = bundle.valenceTracker();
+        this.coActivationTracker = bundle.coActivationTracker();
+        this.suppressionSet = bundle.suppressionSet();
+        this.habituationPenalty = bundle.habituationPenalty();
+        this.prospectiveScheduler = bundle.prospectiveScheduler();
+        this.introspector = bundle.introspector();
+        this.lateralEvaluator = bundle.lateralEvaluator();
+        this.wal = bundle.wal();
+        this.hebbianGraph = bundle.hebbianGraph();
+        this.temporalChain = bundle.temporalChain();
+        this.entityGraph = bundle.entityGraph();
+        this.hyperEntityGraph = bundle.hyperEntityGraph();
+        this.graphFacade = bundle.graphFacade();
         this.dimensions = builder.dimensions;
         this.persistenceMode = builder.persistenceMode;
         this.persistencePath = builder.persistencePath;
         this.circadianPolicy = builder.circadianPolicy;
         this.profileConfig = builder.profileConfig;
+        this.namespaceManager = bundle.namespaceManager();
+        this.idGenerator = bundle.idGenerator();
+        this.checkpointDaemon = bundle.checkpointDaemon();
+        this.daemonSupervisor = bundle.daemonSupervisor();
+        this.bm25Index = bundle.bm25Index();
         this.chunker = builder.chunker;
-
-        if (builder.embeddingProvider == null) {
-            throw new SpectorValidationException(ErrorCode.ARGUMENT_NULL,
-                    "embeddingProvider is required");
-        }
-        EmbeddingProvider embeddingProvider = builder.embeddingProvider;
-        this.embeddingProvider = embeddingProvider;
-        this.parallelPipeline = new ParallelEmbeddingPipeline(embeddingProvider);
-        this.embedConfig = new EmbedConfig(builder.embedBatchSize, 3);
-
-        boolean isDisk = persistenceMode == MemoryPersistenceMode.DISK;
-
-        // ── Resolve persistence path ──
-        Path basePath;
-        if (isDisk && builder.persistencePath != null) {
-            basePath = builder.persistencePath;
-        } else if (isDisk) {
-            basePath = Path.of(System.getProperty("java.io.tmpdir"),
-                    "spector-memory-" + ProcessHandle.current().pid());
-            log.warn("DISK persistence mode with no explicit path — using temp directory: {}", basePath);
-        } else {
-            basePath = null;
-        }
-
-        // ── Quantizer ──
-        if (builder.quantizer != null) {
-            this.quantizer = builder.quantizer;
-        } else {
-            float[] defaultMins = new float[builder.dimensions];
-            float[] defaultMaxs = new float[builder.dimensions];
-            java.util.Arrays.fill(defaultMins, -1.0f);
-            java.util.Arrays.fill(defaultMaxs, 1.0f);
-            this.quantizer = ScalarQuantizer.fromBounds(builder.dimensions, defaultMins, defaultMaxs);
-        }
-
-        // ── Auto-migrate legacy layout ──
-        if (isDisk && basePath != null) {
-            PartitionLayoutMigrator.migrate(basePath);
-        }
-
-        // ── Namespace Manager ──
-        if (isDisk && basePath != null) {
-            this.namespaceManager = new SpectorNamespaceManager(basePath);
-            log.info("NamespaceManager initialized: {} namespaces discovered", namespaceManager.count());
-        } else {
-            this.namespaceManager = null;
-        }
-
-        // ── Partition layout ──
-        int quantizedVecBytes = builder.dimensions;
-
-        Path resolvedPartitionDir = null;
-        if (isDisk && basePath != null) {
-            try {
-                java.nio.file.Files.createDirectories(StorageLayout.runtimeDir(basePath));
-                java.nio.file.Files.createDirectories(StorageLayout.partitionsDir(basePath));
-                resolvedPartitionDir = PartitionManager.discoverOrCreatePartition(basePath);
-                log.info("Active partition: {}", resolvedPartitionDir.getFileName());
-            } catch (java.io.IOException e) {
-                log.error("Failed to initialize partition layout: {}", e.getMessage(), e);
-            }
-        }
-
-        // ── Tier stores ──
-        TierRouter tierRouter;
-        WorkingMemoryStore workingStore;
-        if (isDisk && builder.persistWorkingMemory && basePath != null) {
-            workingStore = new WorkingMemoryStore(quantizedVecBytes, builder.workingCapacity,
-                    StorageLayout.workingMem(basePath));
-        } else {
-            workingStore = new WorkingMemoryStore(quantizedVecBytes, builder.workingCapacity);
-        }
-
-        if (isDisk && basePath != null && resolvedPartitionDir != null) {
-            EpisodicMemoryStore episodicStore = new EpisodicMemoryStore(
-                    StorageLayout.episodicMem(resolvedPartitionDir),
-                    quantizedVecBytes, builder.episodicPartitionCapacity);
-            ProceduralMemoryStore proceduralStore = new ProceduralMemoryStore(
-                    quantizedVecBytes, builder.proceduralCapacity,
-                    StorageLayout.proceduralMem(resolvedPartitionDir));
-            SemanticMemoryStore semanticStore = new SemanticMemoryStore(
-                    quantizedVecBytes, builder.semanticCapacity,
-                    StorageLayout.semanticMem(resolvedPartitionDir));
-            tierRouter = new TierRouter(workingStore, episodicStore, semanticStore, proceduralStore);
-        } else {
-            EpisodicMemoryStore episodicStore = new EpisodicMemoryStore(
-                    quantizedVecBytes, builder.episodicPartitionCapacity);
-            ProceduralMemoryStore proceduralStore = new ProceduralMemoryStore(
-                    quantizedVecBytes, builder.proceduralCapacity);
-            SemanticMemoryStore semanticStore = new SemanticMemoryStore(
-                    quantizedVecBytes, builder.semanticCapacity);
-            tierRouter = new TierRouter(workingStore, episodicStore, semanticStore, proceduralStore);
-        }
-
-        // ── Memory Index ──
-        if (isDisk && basePath != null) {
-            Path runtimeIndex = StorageLayout.indexMidxRuntime(basePath);
-            Path legacyIndex = StorageLayout.legacyIndex(basePath);
-            // V3: load from runtime/, fallback to V2 partition location, then legacy
-            if (java.nio.file.Files.exists(runtimeIndex)) {
-                this.index = MemoryIndex.load(runtimeIndex);
-            } else if (resolvedPartitionDir != null
-                    && java.nio.file.Files.exists(StorageLayout.indexMidx(resolvedPartitionDir))) {
-                this.index = MemoryIndex.load(StorageLayout.indexMidx(resolvedPartitionDir));
-                log.info("Loaded V2 index from partition — will save to runtime/ on next close");
-            } else if (java.nio.file.Files.exists(legacyIndex)) {
-                this.index = MemoryIndex.load(legacyIndex);
-                log.info("Loaded legacy memory-index.mem — will save to runtime/ on next close");
-            } else {
-                this.index = new MemoryIndex();
-            }
-        } else {
-            this.index = new MemoryIndex();
-        }
-
-        // ── WAL ──
-        if (isDisk && basePath != null) {
-            this.wal = new MemoryWal(StorageLayout.walDir(basePath));
-        } else {
-            this.wal = new MemoryWal();
-        }
-
-        // ── Biological Subsystems ──
-        SurpriseDetector surpriseDetector = new SurpriseDetector(builder.surpriseWarmup);
-        IcnuWeights icnuWeights = builder.icnuWeights != null ? builder.icnuWeights : IcnuWeights.DEFAULT;
-        FlashbulbPolicy flashbulbPolicy = new FlashbulbPolicy(builder.flashbulbThreshold);
-        this.valenceTracker = new ValenceTracker(builder.valenceLearningRate);
-
-        if (isDisk && basePath != null) {
-            this.coActivationTracker = CoActivationTracker.load(
-                    StorageLayout.coactivationTracker(basePath), 10_000, 20_000);
-        } else {
-            this.coActivationTracker = new CoActivationTracker();
-        }
-        this.suppressionSet = new SuppressionSet();
-        this.habituationPenalty = new HabituationPenalty(0.2f, builder.inhibitionTtlMs, builder.inhibitionFloor);
-        this.prospectiveScheduler = new ProspectiveScheduler();
-        this.introspector = new MemoryIntrospector(coActivationTracker);
-        this.lateralEvaluator = new LateralEvaluator();
-
-        ReflectDaemon reflectDaemon = new ReflectDaemon(
-                builder.circadianPolicy,
-                builder.dimensions > 0 ? new CentroidRouter(builder.dimensions) : null,
-                builder.textGenerationProvider,
-                embeddingProvider,
-                5, // minClusterSize
-                builder.pinSourceEpisodes,
-                builder.pinnedQuota);
-
-        // ── 3-Layer Cognitive Graph ──
-        int graphCapacity = builder.hebbianGraphCapacity > 0
-                ? builder.hebbianGraphCapacity : builder.episodicPartitionCapacity;
-
-        if (isDisk && basePath != null) {
-            Path runtimeGraph = StorageLayout.hebbianGraphRuntime(basePath);
-            // Fallback: V2 partition location, then legacy flat
-            Path legacyGraph = basePath.resolve(StorageLayout.FILE_HEBBIAN);
-            Path v2Graph = resolvedPartitionDir != null
-                    ? StorageLayout.hebbianGraph(resolvedPartitionDir) : null;
-            Path loadFrom = java.nio.file.Files.exists(runtimeGraph) ? runtimeGraph
-                    : (v2Graph != null && java.nio.file.Files.exists(v2Graph)) ? v2Graph
-                    : legacyGraph;
-            this.hebbianGraph = HebbianGraphCsr.load(loadFrom, graphCapacity,
-                    builder.hebbianMaxDegree, builder.edgeImportance);
-        } else {
-            this.hebbianGraph = new HebbianGraphCsr(graphCapacity);
-        }
-
-        int temporalCapacity = builder.temporalChainCapacity > 0
-                ? builder.temporalChainCapacity : graphCapacity;
-        if (isDisk && basePath != null) {
-            Path runtimeChain = StorageLayout.temporalChainRuntime(basePath);
-            Path legacyChain = basePath.resolve(StorageLayout.FILE_TEMPORAL);
-            Path v2Chain = resolvedPartitionDir != null
-                    ? StorageLayout.temporalChain(resolvedPartitionDir) : null;
-            Path loadFrom = java.nio.file.Files.exists(runtimeChain) ? runtimeChain
-                    : (v2Chain != null && java.nio.file.Files.exists(v2Chain)) ? v2Chain
-                    : legacyChain;
-            this.temporalChain = TemporalChain.load(loadFrom, temporalCapacity);
-        } else {
-            this.temporalChain = new TemporalChain(temporalCapacity);
-        }
-
-        EntityExtractor entityExtractor;
-        if (builder.entityExtractionMode == EntityExtractionMode.LLM
-                && builder.textGenerationProvider != null) {
-            entityExtractor = new LlmEntityExtractor(
-                    builder.textGenerationProvider,
-                    builder.maxEntitiesPerMemory, builder.maxRelationsPerMemory,
-                    builder.llmGenerationOptions);
-        } else if (builder.entityExtractionMode == EntityExtractionMode.CUSTOM
-                && builder.entityExtractor != null) {
-            entityExtractor = builder.entityExtractor;
-        } else {
-            entityExtractor = NoOpEntityExtractor.INSTANCE;
-        }
-
-        boolean entityEnabled = builder.entityExtractionMode != EntityExtractionMode.NONE;
-        if (entityEnabled) {
-            int entityCap = builder.entityGraphCapacity;
-            int edgeCap = entityCap * builder.entityMaxDegree;
-            if (isDisk && basePath != null) {
-                Path runtimeEntity = StorageLayout.entityGraphRuntime(basePath);
-                Path legacyEntity = basePath.resolve(StorageLayout.FILE_ENTITY);
-                Path v2Entity = resolvedPartitionDir != null
-                        ? StorageLayout.entityGraph(resolvedPartitionDir) : null;
-
-                if (java.nio.file.Files.exists(runtimeEntity)) {
-                    // Runtime path exists — use mmap constructor (EGMM) or load (EGPH legacy)
-                    this.entityGraph = EntityGraph.load(runtimeEntity, entityCap, edgeCap);
-                } else if (v2Entity != null && java.nio.file.Files.exists(v2Entity)) {
-                    // V2 legacy path — load as heap (EGPH), will be saved as mmap on next checkpoint
-                    this.entityGraph = EntityGraph.load(v2Entity, entityCap, edgeCap);
-                } else if (java.nio.file.Files.exists(legacyEntity)) {
-                    this.entityGraph = EntityGraph.load(legacyEntity, entityCap, edgeCap);
-                } else {
-                    // No existing file — create fresh mmap-backed graph at runtime path
-                    this.entityGraph = new EntityGraph(runtimeEntity, entityCap, edgeCap,
-                            builder.entityMaxDegree, builder.edgeImportance);
-                }
-            } else {
-                this.entityGraph = new EntityGraph(entityCap, edgeCap,
-                        builder.entityMaxDegree, builder.edgeImportance);
-            }
-        } else {
-            this.entityGraph = null;
-        }
-
-        // ── HyperEntityGraph (multi-entity relationship layer) ──
-        if (entityEnabled && builder.hyperEntityGraphEnabled) {
-            int hyperCap = builder.entityGraphCapacity;
-            int hyperEdgeCap = hyperCap * 2;
-            if (isDisk && basePath != null) {
-                Path hyperPath = StorageLayout.hyperEntityGraphRuntime(basePath);
-                if (java.nio.file.Files.exists(hyperPath)) {
-                    this.hyperEntityGraph = HyperEntityGraph.load(hyperPath, hyperCap, hyperEdgeCap);
-                } else {
-                    this.hyperEntityGraph = new HyperEntityGraph(hyperCap, hyperEdgeCap);
-                }
-            } else {
-                this.hyperEntityGraph = new HyperEntityGraph(hyperCap, hyperEdgeCap);
-            }
-        } else {
-            this.hyperEntityGraph = null;
-        }
-
-        // ── BM25 Text Search ──
-        MemoryBM25Index bm25Index;
-        TextDataStore textDataStore;
-        int activePartitionIndex;
-        if (isDisk && basePath != null && resolvedPartitionDir != null) {
-            textDataStore = new TextDataStore(StorageLayout.textDat(resolvedPartitionDir), builder.dataEncryptor);
-            // Call readAll() to establish the mmap'd segment for off-heap text reads.
-            var textEntries = textDataStore.readAll();
-            // P0: wire TextDataStore into MemoryIndex for off-heap text() resolution
-            index.setTextDataStore(textDataStore);
-
-            // P1: Try loading pre-built BM25 binary index (instant load vs O(n) rebuild)
-            java.nio.file.Path bm25Path = StorageLayout.bm25BidxRuntime(basePath);
-            // Fallback: V2 partition location
-            if (!java.nio.file.Files.exists(bm25Path) && resolvedPartitionDir != null) {
-                java.nio.file.Path v2Bm25 = StorageLayout.bm25Bidx(resolvedPartitionDir);
-                if (java.nio.file.Files.exists(v2Bm25)) bm25Path = v2Bm25;
-            }
-            BM25Index loadedBm25 = BM25Index.load(bm25Path);
-            bm25Index = new MemoryBM25Index(1);
-            if (loadedBm25 != null) {
-                bm25Index.setPartition(0, loadedBm25);
-                log.info("BM25 loaded from binary index: {} docs", loadedBm25.size());
-            } else {
-                // Fall back to expensive rebuild from text data
-                Map<String, String> allTexts = new java.util.HashMap<>();
-                for (var entry : index.locationMap().entrySet()) {
-                    String text = index.text(entry.getKey());
-                    if (text != null && !text.isEmpty()) {
-                        allTexts.put(entry.getKey(), text);
-                    }
-                }
-                if (!allTexts.isEmpty()) {
-                    bm25Index.rebuildPartition(0, allTexts);
-                    log.info("Rebuilt BM25 index with {} documents from memory index", allTexts.size());
-                    // Save the rebuilt index for instant load on next startup
-                    bm25Index.partition(0).save(bm25Path);
-                }
-            }
-            activePartitionIndex = 0;
-        } else {
-            bm25Index = new MemoryBM25Index(1);
-            textDataStore = null;
-            activePartitionIndex = 0;
-        }
-        this.bm25Index = bm25Index;
-
-        // ── SPLADE Index (auto-created when provider is configured) ──
-        com.spectrayan.spector.memory.cortex.MemorySpladeIndex memorySpladeIndex = null;
-        if (builder.sparseEncodingProvider != null) {
-            memorySpladeIndex = new com.spectrayan.spector.memory.cortex.MemorySpladeIndex(1);
-            log.info("SPLADE index enabled: provider={}", builder.sparseEncodingProvider.modelName());
-        }
-
-        // ── ColBERT Reranker (auto-created when provider is configured) ──
-        ColBERTReranker colbertReranker = null;
-        if (builder.tokenEmbeddingProvider != null) {
-            ColBERTTokenCache tokenCache = new ColBERTTokenCache(
-                    builder.tokenEmbeddingProvider.tokenDimensions(), 10_000);
-            colbertReranker = new ColBERTReranker(builder.tokenEmbeddingProvider, tokenCache);
-            log.info("ColBERT reranker enabled: provider={}, tokenDims={}",
-                    builder.tokenEmbeddingProvider.modelName(),
-                    builder.tokenEmbeddingProvider.tokenDimensions());
-        }
-
-        // ── Ingestion Target ──
-        this.cognitiveTarget = new CognitiveIngestionTarget(
-                quantizer, surpriseDetector, flashbulbPolicy,
-                tierRouter, index, wal, workingStore, builder.icnuWeights,
-                builder.semanticIndex, builder.tagExtractor, true,
-                hebbianGraph, temporalChain, entityExtractor, entityGraph,
-                hyperEntityGraph,
-                bm25Index, textDataStore, activePartitionIndex,
-                memorySpladeIndex, builder.sparseEncodingProvider,
-                builder.dataEncryptor);
-
-        // ── Wire Salience Profile Provider ──
-        if (builder.salienceProfileProvider != null) {
-            SalienceProfile effective = builder.salienceProfileProvider.effectiveProfile();
-            if (effective != null && !effective.isNeutral()) {
-                this.cognitiveTarget.setSalienceProfile(effective);
-                log.info("Salience profile applied: interests={}, disinterests={}, icnuOverride={}",
-                        effective.interests().size(), effective.disinterests().size(),
-                        effective.hasIcnuOverride());
-            }
-        }
-
-        // ── Partition Manager ──
-        if (isDisk) {
-            this.partitionManager = new PartitionManager(
-                    basePath, quantizedVecBytes, builder.semanticCapacity,
-                    builder.episodicPartitionCapacity, builder.proceduralCapacity,
-                    tierRouter, resolvedPartitionDir,
-                    index, hebbianGraph, temporalChain, cognitiveTarget);
-            cognitiveTarget.setPartitionRollCallback(partitionManager::rollPartition);
-        } else {
-            this.partitionManager = new PartitionManager(
-                    null, quantizedVecBytes, builder.semanticCapacity,
-                    builder.episodicPartitionCapacity, builder.proceduralCapacity,
-                    tierRouter, null,
-                    index, hebbianGraph, temporalChain, cognitiveTarget);
-        }
-
-        // ── Semantic Recall Strategy + HNSW Rebuild ──
-        SemanticRecallStrategy semanticStrategy = null;
-        if (builder.semanticIndex != null && tierRouter.semantic() != null) {
-            semanticStrategy = new SemanticRecallStrategy(builder.semanticIndex, tierRouter.semantic(), index);
-            rebuildHnswIfNeeded(builder, tierRouter);
-        }
-
-        // ── Recall Pipeline ──
-        this.recallPipeline = new RecallPipeline(
-                embeddingProvider, tierRouter, index,
-                suppressionSet, habituationPenalty, prospectiveScheduler, wal,
-                quantizer.mins(), quantizer.scales(), semanticStrategy,
-                null, hebbianGraph, temporalChain, entityGraph, entityExtractor,
-                builder.graphScoringPolicy, bm25Index,
-                memorySpladeIndex, builder.sparseEncodingProvider, colbertReranker);
-
-        recallPipeline.addListener(new LtpReconsolidationListener(index, tierRouter, wal));
-        recallPipeline.addListener(new HebbianCoActivationListener(coActivationTracker));
-
-        // ── Extracted Components ──
-        this.importanceEstimator = new ImportanceEstimator(
-                surpriseDetector, flashbulbPolicy, icnuWeights, quantizer);
-
-        this.reflectionOrchestrator = new ReflectionOrchestrator(
-                reflectDaemon, hebbianGraph, temporalChain, entityGraph,
-                hyperEntityGraph, wal, builder.temporalRetentionDays);
-
-        this.reinforcementHandler = new ReinforcementHandler(
-                valenceTracker, hebbianGraph, lateralEvaluator, recallPipeline,
-                wal, builder.twoFactorConfig);
-
-        log.info("SpectorMemory initialized: dimensions={}, model={}, persistence={}, mode={}, " +
-                 "partition={}, quantizer={}, idStrategy={}",
-                dimensions, embeddingProvider.modelName(),
-                basePath != null ? basePath : "in-memory",
-                persistenceMode,
-                resolvedPartitionDir != null ? resolvedPartitionDir.getFileName() : "none",
-                builder.quantizer != null ? "user-provided" : "identity-default",
-                builder.idGenerator != null ? "custom" : builder.idStrategy.name());
-
-        // ── Cognitive Graph Facade ──
-        this.graphFacade = new CognitiveGraphFacade(
-                hebbianGraph, temporalChain, entityGraph, hyperEntityGraph, index);
-
-        // ── ID Generator ── (must be after logging)
-        this.idGenerator = builder.idGenerator != null
-                ? builder.idGenerator
-                : builder.idStrategy.createGenerator();
-
-        // ── Daemon Supervisor + Checkpoint Daemon ── (DISK mode only)
-        if (isDisk && basePath != null && builder.checkpointIntervalSeconds > 0) {
-            Path indexSavePath = resolvedPartitionDir != null
-                    ? StorageLayout.indexMidx(resolvedPartitionDir)
-                    : basePath.resolve(StorageLayout.LEGACY_FILE_INDEX);
-            this.checkpointDaemon = new CheckpointDaemon(
-                    tierRouter, wal,
-                    StorageLayout.checkpointMeta(basePath),
-                    index, indexSavePath,
-                    hebbianGraph, temporalChain, entityGraph,
-                    hyperEntityGraph, coActivationTracker,
-                    resolvedPartitionDir, basePath);
-            this.daemonSupervisor = new DaemonSupervisor("memory");
-            this.daemonSupervisor.schedule(
-                    "checkpoint",
-                    checkpointDaemon::checkpoint,
-                    java.time.Duration.ofSeconds(builder.checkpointIntervalSeconds),
-                    DaemonPolicy.CRITICAL);
-        } else {
-            this.checkpointDaemon = null;
-            this.daemonSupervisor = null;
-        }
+        this.parallelPipeline = bundle.parallelPipeline();
+        this.embedConfig = bundle.embedConfig();
+        this.attachmentProcessor = bundle.attachmentProcessor();
 
         // ── JVM Shutdown Hook ── (DISK mode only)
-        // Guarantees a final flush of all cognitive graphs and subsystems
-        // when the JVM exits, even if the caller never calls close().
-        if (isDisk && basePath != null) {
+        if (persistenceMode == MemoryPersistenceMode.DISK && bundle.basePath() != null) {
             this.shutdownHook = new Thread(() -> {
                 if (closed.compareAndSet(false, true)) {
                     log.info("JVM shutdown hook: flushing SpectorMemory...");
@@ -661,53 +270,6 @@ public final class DefaultSpectorMemory implements SpectorMemory, SpectorMemoryA
             Runtime.getRuntime().addShutdownHook(shutdownHook);
         } else {
             this.shutdownHook = null;
-        }
-
-        // ── Multimodal Attachment Processor ──
-        if (!builder.sensoryExtractors.isEmpty()) {
-            this.attachmentProcessor = new com.spectrayan.spector.memory.pipeline.AttachmentProcessor(
-                    builder.sensoryExtractors, builder.assetStore);
-            log.info("AttachmentProcessor initialized with {} extractors", builder.sensoryExtractors.size());
-        } else {
-            this.attachmentProcessor = null;
-        }
-    }
-
-    /**
-     * Rebuilds HNSW index from persisted semantic store if it has data but HNSW is empty.
-     */
-    private void rebuildHnswIfNeeded(SpectorMemoryBuilder builder, TierRouter tierRouter) {
-        var semStore = tierRouter.semantic();
-        int storeSize = semStore.size();
-        if (storeSize > 0 && builder.semanticIndex.size() == 0) {
-            log.info("Rebuilding HNSW index from {} persisted semantic records...", storeSize);
-            long startMs = System.currentTimeMillis();
-            var seg = semStore.primarySegment();
-            var recLayout = semStore.layout();
-            int stride = recLayout.stride();
-            int vecBytes = recLayout.quantizedVecBytes();
-            long baseOffset = semStore.filePath() != null
-                    ? com.spectrayan.spector.memory.cortex.AbstractTierStore.METADATA_HEADER_BYTES : 0;
-
-            int rebuilt = 0;
-            for (int i = 0; i < storeSize; i++) {
-                long recordOff = baseOffset + (long) i * stride;
-                byte[] quantized = new byte[vecBytes];
-                java.lang.foreign.MemorySegment.copy(
-                        seg, java.lang.foreign.ValueLayout.JAVA_BYTE,
-                        recLayout.vectorOffset(recordOff),
-                        java.lang.foreign.MemorySegment.ofArray(quantized),
-                        java.lang.foreign.ValueLayout.JAVA_BYTE, 0, vecBytes);
-
-                float[] vector = quantizer.decode(quantized);
-                String id = index.findIdByOffset(MemoryType.SEMANTIC, recordOff);
-                if (id != null && !builder.semanticIndex.isReadOnly()) {
-                    builder.semanticIndex.add(id, i, vector);
-                    rebuilt++;
-                }
-            }
-            long elapsed = System.currentTimeMillis() - startMs;
-            log.info("HNSW rebuild complete: {}/{} vectors added in {}ms", rebuilt, storeSize, elapsed);
         }
     }
